requirements.txt
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
instaloader==4.14.1
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
instaloader==4.14.1
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
yt-dlp==2025.3.31
